calcula definitivas de las notas

+++ app/src/main/java/com/example/calcularnotaspm/modelo/Nota_Actividad.java
@@ -94,7 +94,7 @@ public class Nota_Actividad {
         this.idMateria = idMateria;
     }
 
-    public double getValor2() {
+    public float getValor2() {
         return (this.valor*this.porcentaje);
     }
 }
